update unit tests after API module restructuring
